Correct use of cmake and ninja in Build section
There is no such thing as `-H.`. I am pretty sure the intention was to use `-S.` (which is a default actually).

`-H` is display help. I am not even sure how it works with `-H.`

For ninja, just use `-C`. No need to change directories, which could be confusing later.

Also the lines before for getting source code use literal directory `vulkandriver`,
so use that in the build too. There is zero reason to complicate it, and if somebody
uses different directory, they are smart enough to change things accordingly.